+++ Sources/Utils/LogUtils/DebugEx.cs
@@ -119,6 +119,15 @@ public static class DebugEx {
       return "[" + componentHost.GetType().Name + ":"
           + DbgFormatter.TranformPath(componentHost.transform) + "]";
     }
+    if (obj is Vector3) {
+      var vec = (Vector3)obj;
+      return string.Format("[Vector3:{0:0.0###},{1:0.0###},{2:0.0###}]", vec.x, vec.y, vec.z); 
+    }
+    if (obj is Quaternion) {
+      var rot = (Quaternion)obj;
+      return string.Format("[Quaternion:{0:0.0###}, {1:0.0###}, {2:0.0###}, {3:0.0###}]",
+                           rot.x, rot.y, rot.z, rot.w);
+    }
     return obj.ToString();
   }
 }